made theta2 and eps2 real*8 instead of real, since they are part in real*8 computations anyway

+++ src/treecode/lpepcsrc/module_allocation.f90
@@ -47,7 +47,7 @@ module module_allocation
         use module_branching
         implicit none
 
-        real, intent(in) :: theta2
+        real*8, intent(in) :: theta2
 
         integer :: nintest
         integer :: nbaddr ! number of bits in the hashing function
